+++ lib/xnix/container.rb
@@ -35,6 +35,10 @@ module Xnix
       runtime_command(network: "bridge", extra_mounts: [source_cache_mount], command: command)
     end
 
+    def cache_run_command(command)
+      runtime_command(network: "none", extra_mounts: [source_cache_mount], command: command)
+    end
+
     private
 
     def runtime_command(network:, extra_mounts:, command:)