+++ scripts/container.rb
@@ -4,6 +4,7 @@
 require "open3"
 require "pathname"
 require_relative "../lib/xnix/container"
+require_relative "../lib/xnix/buildroot"
 
 PROJECT_ROOT = Pathname.new(__dir__).join("..").realpath
 VERSION = PROJECT_ROOT.join("VERSION").read.strip
@@ -16,6 +17,7 @@ end
 abort "Xnix commands require the Colima Docker context" unless colima_context?
 
 container = Xnix::Container.new(project_root: PROJECT_ROOT.to_s, version: VERSION)
+buildroot = Xnix::Buildroot.new
 
 case ARGV.shift
 when "build"
@@ -28,6 +30,14 @@ when "fetch-sources"
   abort "Usage: ruby scripts/container.rb fetch-sources" unless ARGV.empty?
 
   exec(*container.source_retrieval_command(["ruby", "scripts/fetch_buildroot.rb"]))
+when "configure-system"
+  abort "Usage: ruby scripts/container.rb configure-system" unless ARGV.empty?
+
+  exec(*container.cache_run_command(buildroot.configure_command))
+when "build-system"
+  abort "Usage: ruby scripts/container.rb build-system" unless ARGV.empty?
+
+  exec(*container.cache_run_command(buildroot.build_command))
 else
-  abort "Usage: ruby scripts/container.rb {build|fetch-sources|offline-run COMMAND [ARGUMENT ...]}"
+  abort "Usage: ruby scripts/container.rb {build|build-system|configure-system|fetch-sources|offline-run COMMAND [ARGUMENT ...]}"
 end